requirements.txt
Flask
line-bot-sdk
python-dotenv
gunicorn
firebase-admin
requests
beautifulsoup4
google-genai